feat(pattern): L2b 기존설계 회랑 소프트 바이어스 — 회랑 셀 주입 + w_corridor
매칭되는 기존 설계배관(TB_ROUTE_PATH) 폴리라인을 회랑으로 주입해, 새 경로가 사람 설계를
부드럽게 따라가게 한다(충돌은 여전히 회피). 옵트인(기본 OFF).

[엔진 — corridor 키 64비트화]
- cost.hpp/astar.hpp/multi_route.hpp: CostModel corridor 셋을 unordered_set<int> →
  unordered_set<long long>(occ.lin) 로 확장 → ImplicitOccupancy 64비트 lin 대형격자에서
  키 절단 없음. 회랑은 w_corridor>0 일 때만 동작하므로 골든(w_corridor=0)은 완전 불변.
- test_attract: 회랑 셋 타입 long long 으로 갱신.

[C ABI — 회랑 셀 주입]
- r3d_set_corridor_cells(e, ijk, n): 학습 회랑 셀(ijk 삼중항)을 엔진에 저장(n<=0=초기화).
- R3dEngine.corridor_seed + route_multi_impl/route_multi_into_doc 가 시드를 받아, w_corridor>0
  이면 회랑에 미리 넣고(occ.lin) 깔린 배관 곁 번들링과 함께 사용. route_multi/_progress 전달.

[C# 뷰어 — 기존설계 회랑]
- Native/Engine: r3d_set_corridor_cells / SetCorridorCells 래퍼.
- SceneViewModel: UseDesignCorridor 토글(기본 OFF) + BuildDesignCorridorCells(매칭 기존배관
  폴리라인을 셀 복셀화 ±2) + SetParams w_corridor=½칸. MainWindow 체크박스.
- DbRouteDiag: 헤드리스 A/B(env R3D_CORRIDOR=on) + MatchPipe.

[검증] ctest golden/capi/implicit/attract/corridor 5/5 불변(회랑 OFF 경로 바이트 동일).
A/B(project6 c100 ALL): 회랑 OFF 194/208·6.5s → ON 198/208·46s·총길이 1,458,300.
회랑 ON = +4 성공(설계 채널 유도) · 총길이↑(설계 정합) · ~7× 느림(w_corridor 페널티). 옵트인.

[문서] docs/routing3d_pattern_learning_plan.{docx,pdf} 13.4 절을 'L2b 구현 완료'로 갱신.

+++ cpp/include/routing3d/astar.hpp
@@ -162,7 +162,7 @@ AStarResult astar(const Occ& occ, Cell start, Cell goal, double step_cost = -1.0
 template <class Occ>
 AStarResult astar_weighted(const Occ& occ, Cell start, Cell goal, const RouteParams& params,
                            long long max_expansions = -1, bool collect_visited = false,
-                           const std::unordered_set<int>* corridor = nullptr,
+                           const std::unordered_set<long long>* corridor = nullptr,
                            const std::function<void(long long, double)>* on_progress = nullptr,
                            long long progress_every = 0) {
     auto t0 = detail::Clock::now();
